Added while block
