+++ CommonTesting/Conditionals.swift
@@ -1,0 +1,31 @@
+//
+//  Conditionals.swift
+//  Tracery
+//
+//  Created by Benzi on 14/03/17.
+//  Copyright © 2017 Benzi Ahamed. All rights reserved.
+//
+
+import XCTest
+@testable import Tracery
+
+class Conditionals: XCTestCase {
+    
+    func testIfBlock() {
+        
+//        let line = "[if #1#==#1# then [tag:you] else [tag:me]]#tag#"
+//        print(Lexer.tokens(input: line))
+        
+        Tracery.logLevel = .verbose
+        let t = Tracery.hierarchical {[
+            "number" : [-1,0,1,2,3,4,5,6,7,8,9],
+            ]}
+        
+        t.setCandidateSelector(rule: "number", selector: SequentialSelector())
+        
+        Tracery.maxStackDepth = 30
+        print(t.expand("[while #number#!=9 do #number#,]"))
+        
+    }
+    
+}
